requirements.txt
watchdog>=2.1.9
numpy>=1.22.0
opencv-python>=4.6.0
torch>=1.12.0
torchvision>=0.13.0
transformers>=4.21.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
PySide6>=6.3.0
SQLAlchemy>=1.4.40
pillow>=9.2.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
